Reorganized changeModulation
Took contents of changeModulation and shifted them to configFSK. Added  enableOOK to beginFSK as a way to switch to OOK modulation. The enbleOOK flag is saved within SX127x class in _OOKEnabled. setBitRate was modify to check for allowed bit rate in OOK mode. Actual bitrate math was left untouched as both FSK and OOK follow the same math, except BitFraction has no effect and is set to 0.